增加资源组相关的打包处理

+++ Assets/CatAsset/Runtime/CatAssetComponent.cs
@@ -13,7 +13,7 @@ namespace CatAsset
     {
         public RunMode RunMode = RunMode.PackageOnly;
         public int MaxTaskExuteCount = 10;
-        public int UnloadDelayTime = 5;
+        public float UnloadDelayTime = 5;
         public bool IsEditorMode = true;
         public float EditorModeMaxDelay = 1;
